submit_api/requirements.txt
fastapi
pydantic
uvicorn
python-jose
passlib[bcrypt]
python-multipart
sqlalchemy